Revert back to gpgme-1.17.0
This reverts b46b5db21813b39584eb1046239dfba7bb40571e which bumped
the libqgpgme soname from 7 to 15.
This was built in a side tag but never actually shipped.

With the Python 3.11 rebuild, the soname bump was merged to rawhide
and blocks the compose:
https://pagure.io/releng/failed-composes/issue/3682

Considering the severity of this
and the fact that the dependent KDE packages FTBFS,
this change is now reverted.
It has never been shipped in a compose, so an epoch bump is not necessary.

We need to pass -b3 to %autorelease to produce
gpgme-1.17.0-3.fc37 instead of gpgme-1.17.0-1.fc37.

When the package is updated again, remove that.

Leaving the tests disabled, because i686 still fails:
https://bugzilla.redhat.com/show_bug.cgi?id=2096365